requirements.txt
# Global:
fire==0.1.3
unidecode
protobuf
# STT
google-cloud-speech==0.36.3
pyaudio==0.2.11
six==1.12.0
mock
# charnn
tensorflow==1.13.1
numpy==1.16.2
# TTS
pyttsx3==2.7
